+++ src/main/java/prophet/command/CommandType.java
@@ -1,3 +1,5 @@
+package prophet.command;
+
 public enum CommandType {
     LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, UNKNOWN
 }
