+++ src/main/java/prophet/command/ListCommand.java
@@ -1,3 +1,8 @@
+package prophet.command;
+
+import prophet.storage.Storage;
+import prophet.ui.Ui;
+
 public class ListCommand extends Command {
     /**
      * The ListCommand class represents a command to list all tasks in the chatbot.
