+++ src/main/java/prophet/exception/NoDescriptionException.java
@@ -1,3 +1,5 @@
+package prophet.exception;
+
 public class NoDescriptionException extends ProphetException {
     protected String message;
