+++ src/main/java/prophet/task/DeadlineTask.java
@@ -1,3 +1,5 @@
+package prophet.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
