+++ src/main/java/prophet/task/Task.java
@@ -1,3 +1,5 @@
+package prophet.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
