+++ src/main/java/prophet/task/ToDoTask.java
@@ -1,3 +1,5 @@
+package prophet.task;
+
 public class ToDoTask extends Task {
     protected static final String type = "T | ";
     public ToDoTask(String description) {
